requirements.txt
cffi==1.15.1
cryptography==40.0.2
fastapi==0.68.1
httpx==0.19.0
python-jose==3.3.0
PyJWT==2.8.0
uvicorn==0.15.0
pycparser==2.21